refactored for delivery
