+++ src/components/sidebar.jsx
@@ -13,125 +13,127 @@ import QuickAdd from "./quickadd";
 class Sidebar extends Component {
   render() {
     return (
-      <div className="w-1/5 shadow mt-12 fixed">
-        <div className="inline-flex flex-col space-y-2 items-start justify-start flex-1 h-full px-6 pt-6 pb-12">
-          <QuickAdd></QuickAdd>
-          {/* navlinks */}
-          {/* dashboard */}
-          <Link to="/dash" className="w-full">
-            <div
-              className={
-                "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
-                (this.props.place == 1 ? "bg-purple-100" : "")
-              }
-            >
-              <div className="flex space-x-8 items-center justify-start">
-                <DashboardIcon className="h-5 " />
-                <p
-                  className={
-                    "text-sm font-medium " +
-                    (this.props.place == 1
-                      ? "text-indigo-600 "
-                      : "text-gray-400")
-                  }
-                >
-                  Dashboard
-                </p>
+      <div className="w-1/5 shadow mt-12 fixed h-full">
+        <div className="inline-flex flex-col space-y-2 items-start justify-between flex-1 h-full px-6 pt-6 pb-12">
+          <div>
+            <QuickAdd></QuickAdd>
+            {/* navlinks */}
+            {/* dashboard */}
+            <Link to="/dash" className="w-full">
+              <div
+                className={
+                  "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
+                  (this.props.place == 1 ? "bg-purple-100" : "")
+                }
+              >
+                <div className="flex space-x-8 items-center justify-start">
+                  <DashboardIcon className="h-5 " />
+                  <p
+                    className={
+                      "text-sm font-medium " +
+                      (this.props.place == 1
+                        ? "text-indigo-600 "
+                        : "text-gray-400")
+                    }
+                  >
+                    Dashboard
+                  </p>
+                </div>
               </div>
-            </div>
-          </Link>
-          {/* Products */}
-          <Link to="/products" className="w-full">
-            <div
-              className={
-                "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
-                (this.props.place == 2 ? "bg-purple-100" : "")
-              }
-            >
-              <div className="inline-flex space-x-8 items-center justify-start">
-                <BoxIcon className="h-5 " />
-                <p
-                  className={
-                    "text-sm font-medium " +
-                    (this.props.place == 2
-                      ? "text-indigo-600 "
-                      : "text-gray-400")
-                  }
-                >
-                  Products
-                </p>
+            </Link>
+            {/* Products */}
+            <Link to="/products" className="w-full">
+              <div
+                className={
+                  "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
+                  (this.props.place == 2 ? "bg-purple-100" : "")
+                }
+              >
+                <div className="inline-flex space-x-8 items-center justify-start">
+                  <BoxIcon className="h-5 " />
+                  <p
+                    className={
+                      "text-sm font-medium " +
+                      (this.props.place == 2
+                        ? "text-indigo-600 "
+                        : "text-gray-400")
+                    }
+                  >
+                    Products
+                  </p>
+                </div>
               </div>
-            </div>
-          </Link>
-          {/* billing */}
-          <Link to="/billing" className="w-full">
-            <div
-              className={
-                "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
-                (this.props.place == 3 ? "bg-purple-100" : "")
-              }
-            >
-              <div className="inline-flex space-x-8 items-center justify-start">
-                <FileTextIcon className="h-5 " />
-                <p
-                  className={
-                    "text-sm font-medium " +
-                    (this.props.place == 3
-                      ? "text-indigo-600 "
-                      : "text-gray-400")
-                  }
-                >
-                  Billing
-                </p>
+            </Link>
+            {/* billing */}
+            <Link to="/billing" className="w-full">
+              <div
+                className={
+                  "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
+                  (this.props.place == 3 ? "bg-purple-100" : "")
+                }
+              >
+                <div className="inline-flex space-x-8 items-center justify-start">
+                  <FileTextIcon className="h-5 " />
+                  <p
+                    className={
+                      "text-sm font-medium " +
+                      (this.props.place == 3
+                        ? "text-indigo-600 "
+                        : "text-gray-400")
+                    }
+                  >
+                    Billing
+                  </p>
+                </div>
               </div>
-            </div>
-          </Link>
-          {/* vendors */}
-          <Link to="/vendors" className="w-full">
-            <div
-              className={
-                "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
-                (this.props.place == 4 ? "bg-purple-100" : "")
-              }
-            >
-              <div className="inline-flex space-x-8 items-center justify-start">
-                <UsersIcon className="h-5 " />
-                <p
-                  className={
-                    "text-sm font-medium " +
-                    (this.props.place == 4
-                      ? "text-indigo-600 "
-                      : "text-gray-400")
-                  }
-                >
-                  Vendors
-                </p>
+            </Link>
+            {/* vendors */}
+            <Link to="/vendors" className="w-full">
+              <div
+                className={
+                  "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
+                  (this.props.place == 4 ? "bg-purple-100" : "")
+                }
+              >
+                <div className="inline-flex space-x-8 items-center justify-start">
+                  <UsersIcon className="h-5 " />
+                  <p
+                    className={
+                      "text-sm font-medium " +
+                      (this.props.place == 4
+                        ? "text-indigo-600 "
+                        : "text-gray-400")
+                    }
+                  >
+                    Vendors
+                  </p>
+                </div>
               </div>
-            </div>
-          </Link>
-          {/* delivery */}
-          <Link to="/delivery" className="w-full">
-            <div
-              className={
-                "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
-                (this.props.place == 5 ? "bg-purple-100" : "")
-              }
-            >
-              <div className="inline-flex space-x-8 items-center justify-start">
-                <TruckIcon className="h-5 " />
-                <p
-                  className={
-                    "text-sm font-medium " +
-                    (this.props.place == 5
-                      ? "text-indigo-600 "
-                      : "text-gray-400")
-                  }
-                >
-                  Delivery
-                </p>
+            </Link>
+            {/* delivery */}
+            <Link to="/delivery" className="w-full">
+              <div
+                className={
+                  "inline-flex items-center justify-start w-full h-12 pl-2 pr-16 pt-2 pb-2.5 rounded-lg " +
+                  (this.props.place == 5 ? "bg-purple-100" : "")
+                }
+              >
+                <div className="inline-flex space-x-8 items-center justify-start">
+                  <TruckIcon className="h-5 " />
+                  <p
+                    className={
+                      "text-sm font-medium " +
+                      (this.props.place == 5
+                        ? "text-indigo-600 "
+                        : "text-gray-400")
+                    }
+                  >
+                    Delivery
+                  </p>
+                </div>
               </div>
-            </div>
-          </Link>
+            </Link>
+          </div>
           {/* faqs */}
           <div className="flex flex-col w-full space-y-2.5 items-center justify-center p-6 bg-purple-100 rounded-2xl">
             <p className="w-40 text-lg font-bold text-center text-gray-900">
