+++ src/routes/delivery/delivery.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import "../index.css";
-import Header from "../components/header";
-import Sidebar from "../components/sidebar";
-import DeliveryMainCard from "../components/delivery/maincard";
+import "../../index.css";
+import Header from "../../components/header";
+import Sidebar from "../../components/sidebar";
+import DeliveryMainCard from "../../components/delivery/maincard";
 
 const Delivery = () => {
   return (
@@ -13,10 +13,10 @@ const Delivery = () => {
       {/* main content container */}
       <div className="mt-auto w-4/5 p-10 bg-gray-50 ml-auto">
         <div className="flex flex-col">
-        <DeliveryMainCard />
-        <DeliveryMainCard />
-        <DeliveryMainCard />
-        <DeliveryMainCard />
+          <DeliveryMainCard />
+          <DeliveryMainCard />
+          <DeliveryMainCard />
+          <DeliveryMainCard />
         </div>
       </div>
     </div>
